Add a templated port function. Clean up model API
In this commit I try moving more of the enq and deq smarts
up into the superclass SWmodel. Subclasses now implement a tick function
which always advances that model locally a single cycle. Model writers act
on ports who have a single member reflecting the data type that
parameterized it. Instead it is the duty of SWModel to check that
all input tokens are valid before calling tick